requirements.txt
pandas 
numpy 
streamlit==1.22.0 
plotly_express==0.4.1   
datetime
google-auth
google-cloud-bigquery==3.9.0
